requirements.txt
fastapi
uvicorn
requests
python-dotenv
sqlalchemy
pydantic
streamlit
pandas